backend/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.13"
asgiref==3.9.1 ; python_version >= "3.13"
attrs==25.3.0 ; python_version >= "3.13"
cachetools==5.5.2 ; python_version >= "3.13"
certifi==2025.7.14 ; python_version >= "3.13"
charset-normalizer==3.4.2 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
django-cors-headers==4.7.0 ; python_version >= "3.13"
django==5.2.4 ; python_version >= "3.13"
djangorestframework==3.16.0 ; python_version >= "3.13"
drf-spectacular==0.28.0 ; python_version >= "3.13"
google-ai-generativelanguage==0.6.15 ; python_version >= "3.13"
google-api-core==2.25.1 ; python_version >= "3.13"
google-api-python-client==2.177.0 ; python_version >= "3.13"
google-auth-httplib2==0.2.0 ; python_version >= "3.13"
google-auth==2.40.3 ; python_version >= "3.13"
google-generativeai==0.8.5 ; python_version >= "3.13"
googleapis-common-protos==1.70.0 ; python_version >= "3.13"
grpcio-status==1.71.2 ; python_version >= "3.13"
grpcio==1.74.0 ; python_version >= "3.13"
httplib2==0.22.0 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
inflection==0.5.1 ; python_version >= "3.13"
jsonpickle==4.1.1 ; python_version >= "3.13"
jsonschema-specifications==2025.4.1 ; python_version >= "3.13"
jsonschema==4.25.0 ; python_version >= "3.13"
proto-plus==1.26.1 ; python_version >= "3.13"
protobuf==5.29.5 ; python_version >= "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pydantic-core==2.33.2 ; python_version >= "3.13"
pydantic==2.11.7 ; python_version >= "3.13"
pyparsing==3.2.3 ; python_version >= "3.13"
pyyaml==6.0.2 ; python_version >= "3.13"
referencing==0.36.2 ; python_version >= "3.13"
requests==2.32.4 ; python_version >= "3.13"
rpds-py==0.26.0 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
sqlparse==0.5.3 ; python_version >= "3.13"
tqdm==4.67.1 ; python_version >= "3.13"
typing-extensions==4.14.1 ; python_version >= "3.13"
typing-inspection==0.4.1 ; python_version >= "3.13"
tzdata==2025.2 ; python_version >= "3.13" and sys_platform == "win32"
uritemplate==4.2.0 ; python_version >= "3.13"
urllib3==2.5.0 ; python_version >= "3.13"
